% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation.R
\name{change_BY}
\alias{change_BY}
\title{Title}
\usage{
change_BY(BY = 2020, inflation_table = jic)
}
\arguments{
\item{BY:}{Base Year}

\item{inflation_table:}{default is to use latest jic}
}
\value{
data frame
}
\description{
Title
}
