% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation.R
\name{inflation}
\alias{inflation}
\title{Master Inflation Function}
\usage{
inflation(
  Index,
  FromYR,
  ToYR,
  Cost,
  from_type = "BY",
  to_type = "BY",
  inflation_table = NULL
)
}
\arguments{
\item{Index}{name of inflation indices}

\item{inflation_table}{}
}
\value{

}
\description{
Master Inflation Function
}
\examples{
nToYR = ToYR
}
