% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{lc_midpoint}
\alias{lc_midpoint}
\title{Title}
\usage{
lc_midpoint(f, l, LC = 1, b = log(LC))
}
\arguments{
\item{f}{first unit of lot}

\item{l}{last unit of lot}

\item{LC}{learning curve parameter (optional)}

\item{b}{learning curve coefficient (optional)}
}
\value{

}
\description{
Title
}
