% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn_curve}
\alias{learn_curve}
\title{Learning Curve Function}
\usage{
learn_curve(T1, LC, RC, Qty, rate = 1)
}
\arguments{
\item{T1}{numeric value}

\item{LC}{decimal value}

\item{rate}{}
}
\value{

}
\description{
Learning Curve Function
}
