% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasing.R
\name{phase_rayleigh}
\alias{phase_rayleigh}
\title{Rayleigh Distribution}
\usage{
phase_rayleigh(StartDate, Duration, Cost, delta = 0.03, return_df = T)
}
\arguments{
\item{df}{}
}
\value{

}
\description{
Rayleigh Distribution
}
